Moving default values into separate files.
Make DefaultCarrierConfigService easier to maintain by moving the values
for each network into XML files. It's important to note that this uses
assets and not resources. The file for a particular network may depend
on more than just the MCC+MNC in the future.

Note on the XML document format. This is set by
PersistableBundle.restoreFromXml, so there is no option to change it in
this CL.

Bug: 21618018
Change-Id: Iecfe7009c2aa0f7a9d5814d1f355dfe4190ce912