Migrate 'csp_enabled' overrides.
Currently the variable is only used in *DEVICE* overlays for specific
networks, but we are dropping the device-overlay so that this simply
applies to those networks. (e.g. there were shamu overlays changing this
variable on the same networks modified in this CL.

This variable controls whether SIM cards are allowed to disable manual
carrier selection, so it makes more sense for it to be managed as
carrier config.

Change-Id: I9a58017808eac9ff28f6ef4d6a5c6751f4773091