Move Telephony overlay config to CarrierConfig
The variable disable_cdma_activation_code_bool is overridden on shamu
for three specific VZW networks, but the overrides have been moved to
vendor/moto/shamu. They can be deleted from here.

Note: 310304 was supposed to be 310004 (corrected in vendor/moto/shamu)
and 20404 was only supposed to apply for a specific gid1.

Bug: 22229196
Bug: 21618018
Change-Id: I7b4e78480d054daf595f856dab3ffa18979f8f54