Add IMSI and SPN regexp filtering for MVNO support
The filtering criterias for selecting carrier configurations are
extended to cover IMSI. The reason for this to support MVNO
configuration for IMSI ranges (i.e. when the only way to
identify a MVNO is by the IMSI)

The xml arguments for IMSI and SPN can be designed as regular
expressions in order to add flexibility.

Some operators do not have SPN defined on there SIM card, but the
MVNOs in their network might have. This is indicated in the xml
by setting the spn argument to "null".

Bug: 32398239
Test: Verified on target with MVNO configs.
Change-Id: Ib4fd87fe484a50fd808fbe9843061ab894703531