Use default RSRP thresholds for Vodafone NL.
This CL moves the Verizon specific RSRP thresholds under VerizonGID for dual IMSI
case so it doesn't impact Vodafone NL.

The history on how this became incorrect for Vodafone NL is below

As per comments in go/aog/209623, Vodafone NL was not the target carrier but only
its MVNO verizon_us (dual-IMSI). We would like to remove this. go/aog/209623
introduced config resource `config_lteDbmThresholds`, 20404 override
= [(-140), -115, ...], default = [(-140), -128, ...] to replace
`config_LTE_RSRP_threshold_type`, 20404 override = 0, default = 1 and this CL is
a no-op.

Why `config_LTE_RSRP_threshold_type` was 0 (instead of default 1) for 20404?
It's by ag/579171. The CL only changed 20404 and 311480. The related bug
(b/17891524#comment10, the bug itself is about orange_fr) pointed to b/17335556
where *Verizon* signal issue was complained.

So, we believe the 20404 customization was for verizon_us instead of vodafone_nl
or any other MVNO.

Test: make
Bug: 74563316
Change-Id: Ibf0d044d45ef1b8b27edc4ac12b8c08281c2f24e