Move the app to /system_ext
As mentioned in go/partitions-in-r, we plan to forbid hidden API in
product partition.
So this app should be moved to /system_ext because it uses hidden API.
It is not permanent, so they can be moved back to /product due to
needs if there isn't hidden usage anymore.

Bug: 137908189
Test: m
Test: check whether device boots successfully
Exempt-From-Owner-Approval: cherry-pick from internal

Change-Id: I1a5e10c08f7c5574f5270d0aeb79ddbf610f9e2a
Merged-In: I1a5e10c08f7c5574f5270d0aeb79ddbf610f9e2a