Add support to fetch carrier config for no SIM case.
When there is no SIM card, CarrierConfigManager returns the code
default values to the clients. This is a problem for certain carriers
if they wanted a different behavior for no SIM case.

The CarrierConfigLoader would query the default carrier app with null
carrier ID when the SIM card is absent. This change handles the
null carrier ID and reads the config from a no SIM asset where the OEMs
could override CarrierConfigManager default.

Bug: 142090499
Test: Manually verified by removing SIM card
      Manually verified by rebooting w/o SIM card
Change-Id: I5d22d202d50963657aabf544a10d88c258d670e8