Change ECC list config for carrier ID 1454.
Ensure for this carrier that calls are determined to be an emergency call
based ONLY on the emergency numbers known by the SIM associated with
this carrier.  This is applicable in multi-sim scenarios where the other
SIM might know of more emergency numbers.

Test: Added unit tests for this config option.
Bug: 203947159
Change-Id: If47f76ed9e034ed73c84cb0d92f2d31ffb1fbbf9
(cherry picked from commit 8c973eea8830b3065a05de5726904fa5611496cf)