Used single TCP configuration for devices
Starting from Android 13, the TCP buffer size is fixed after boot up,
and should never be changed based on carriers or the network types.
The value should be configured appropriately based on the device's
memory and performance. It is recommended to use lower values if the
device has low memory or doesn't support high-speed network such like
LTE, NR, or Wifi.

Bug: 213596972
Test: Manual
Change-Id: I4236cbf8ff0d7d8bc9fde69c7b63d8c92dd9e99a